+++ app.js
@@ -1,5 +1,13 @@
-var app = require('./search');
+const express = require('express');
+const app = express();
+const handlers = require('./handlers');
 
-app.listen(8080, () => console.log('server started port 8080'));
+app.get('/', function(req, res) {
+    handlers.search(req.query)
+        .fork(
+            err => res.status(err.status || 500).send(err.message),
+            r => res.json(r)
+        )
+});
 
 module.exports = app;
