+++ handlers/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    search: require('./search').search
+}
