--- search.js
@@ -1,13 +1,0 @@
-const express = require('express');
-const app = express();
-const searchHandler = require('./searchHandler');
-
-app.get('/', function(req, res) {
-    searchHandler(req.query)
-        .fork(
-            err => res.status(err.status || 500).send(err.message),
-            r => res.json(r)
-        )
-});
-
-module.exports = app;
