+++ start.js
@@ -1,0 +1,5 @@
+var app = require('./app');
+
+app.listen(8080, () => console.log('server started port 8080'));
+
+module.exports = app;
